Small refactoring
- new to_list utilities
- split test_stores into fsspec-based and argo-gdac stores
